Refactor various mocks into their own headers
This should speed up compilation for tests that use only these mocks.
