+++ test/mocks/http/stream_decoder.cc
@@ -1,0 +1,10 @@
+#include "test/mocks/http/stream_decoder.h"
+
+namespace Envoy {
+namespace Http {
+
+MockStreamDecoder::MockStreamDecoder() {}
+MockStreamDecoder::~MockStreamDecoder() {}
+
+} // namespace Http
+} // namespace Envoy
