+++ test/mocks/upstream/load_balancer_context.cc
@@ -1,0 +1,11 @@
+#include "test/mocks/upstream/load_balancer_context.h"
+
+namespace Envoy {
+namespace Upstream {
+
+MockLoadBalancerContext::MockLoadBalancerContext() = default;
+
+MockLoadBalancerContext::~MockLoadBalancerContext() = default;
+
+} // namespace Upstream
+} // namespace Envoy
